+++ lib/core/utils/my_logs.dart
@@ -1,8 +1,9 @@
 // ignore_for_file: avoid_print
 
 
+import 'package:accredit/core/settings.dart';
 import 'package:flutter/material.dart';
-import 'package:lazycopy/core/settings.dart';
+
 import 'package:logging/logging.dart';